OpenWRT: MVLAN-PDN IPv6CT additional fixes
MVLAN-PDN IPv6CT additional fixes to support IPA
HW path

+++ ipacm/src/IPACM_Conntrack_NATApp.cpp
@@ -2664,18 +2664,53 @@ int NatBase::AddTable(const uint32_t v6_prefix[2])
 	return 0;
 }
 
-int NatBase::DeleteTable(const IpAddress& wan_addr)
+int NatBase::DeleteTable(const uint32_t v6_prefix[2],int num_v6_vlan_pdns)
 {
 	IPACMDBG_H("\n");
-	int ret = m_proxy.DeleteTable();
-	if (ret)
+
+	for (int cnt = 0; cnt < m_maxEntries; ++cnt)
 	{
-		IPACMERR("unable to delete the table Error: %d\n", ret);;
-		return ret;
+		NatEntryBase& entry = m_cache[cnt];
+		if (entry.Valid())
+		{
+			uint64_t src_ipv6_msb;
+
+			if (entry.m_direction == NatEntryBase::DirectionOutbound || entry.m_direction == NatEntryBase::DirectionUnknown)
+			{
+				src_ipv6_msb = ((Ipv6IpAddress &)entry.GetClientIp()).GetMsb();
+			}
+			else if (entry.m_direction == NatEntryBase::DirectionInbound)
+			{
+				src_ipv6_msb = ((Ipv6IpAddress &)entry.GetTargetIp()).GetMsb();
+			}
+
+			if(ipv6prefixmatch(src_ipv6_msb, v6_prefix))
+			{
+				if (m_proxy.DelEntry(entry))
+				{
+					IPACMERR("unable to delete the rule delete from cache\n");
+					entry.Clear();
+					--m_curCnt;
+					continue;
+				}
+
+				entry.DebugDump("On wan-iface reset added below rule successfully\n");
+			}
+		}
 	}
 
-	m_previousWanAddress = wan_addr;
-	Reset();
+	if(num_v6_vlan_pdns == 0)
+	{
+		int ret = m_proxy.DeleteTable();
+		if (ret)
+		{
+			IPACMERR("unable to delete the table Error: %d\n", ret);;
+			return ret;
+		}
+ 
+		Reset();
+	}
+
 	IPACMDBG_H("return\n");
 	return 0;
 }
@@ -2900,6 +2935,62 @@ void NatBase::FlushTempEntries(const IpAddress& clientIp, bool isAdd, bool isDum
 				IPACMERR("unable to add temp entry: %d\n", ret);
 				continue;
 			}
+			IPACMDBG_H("Successfully flushed the entry\n");
+		}
+
+		curr.Clear();
+	}
+	IPACMDBG_H("return\n");
+}
+
+void NatBase::FlushAndCacheVlanTempEntries_v6(uint32_t* ipv6_addr, bool isAdd, bool isDummy, bool isStaClientIp)
+{
+	IPACMDBG_H("\n");
+	const IpAddress& clientIp = Ipv6IpAddress(ipv6_addr, false);
+	clientIp.DebugDump("Flush temp entries for");
+
+	int valid = 0;
+
+	for (int cnt = 0; cnt < MAX_TEMP_ENTRIES; ++cnt)
+	{
+		NatEntryBase& curr = m_temp[cnt];
+		if (!curr.UpdateDirection(clientIp, isStaClientIp))
+		{
+			continue;
+		}
+		curr.DebugDump((isAdd) ? "Add temp entry to cache" : "Delete temp entry");
+
+		uint64_t src_ipv6_msb;
+
+		if (curr.m_direction == NatEntryBase::DirectionOutbound || curr.m_direction == NatEntryBase::DirectionUnknown)
+		{
+			src_ipv6_msb = ((Ipv6IpAddress &)curr.GetClientIp()).GetMsb();
+		}
+		else if (curr.m_direction == NatEntryBase::DirectionInbound)
+		{
+			src_ipv6_msb = ((Ipv6IpAddress &)curr.GetTargetIp()).GetMsb();
+		}
+
+		if(((src_ipv6_msb >> 32) == ipv6_addr[0]) && ((src_ipv6_msb & 0x00000000FFFFFFFF) == ipv6_addr[1]))
+		{
+			valid = 1;
+		}
+
+		if (isAdd && valid)
+		{
+			if (isDummy)
+			{
+				IPACMDBG("setting to dummy\n");
+				curr.m_isDummy = true;
+			}
+
+			int ret = AddEntry(curr);
+			if (ret)
+			{
+				IPACMERR("unable to add temp entry: %d\n", ret);
+				continue;
+			}
+
 			IPACMDBG_H("Successfully flushed the entry\n");
 		}
